switch to upstream recipe; enable cuda

+++ recipe/build-pyarrow.sh
@@ -20,21 +20,27 @@ set -x
 # Build dependencies
 export ARROW_HOME=$PREFIX
 export PARQUET_HOME=$PREFIX
-export PYARROW_BUNDLE_ARROW_CPP_HEADERS=0
 export SETUPTOOLS_SCM_PRETEND_VERSION=$PKG_VERSION
 export PYARROW_BUILD_TYPE=release
+export PYARROW_BUNDLE_ARROW_CPP_HEADERS=0
 export PYARROW_WITH_DATASET=1
 export PYARROW_WITH_FLIGHT=1
+# Disable Gandiva
+export PYARROW_WITH_GANDIVA=0
 export PYARROW_WITH_HDFS=1
 export PYARROW_WITH_ORC=1
-export PYARROW_WITH_PLASMA=1
 export PYARROW_WITH_PARQUET=1
-export PYARROW_WITH_ORC=1
-export PYARROW_WITH_S3=1
-export PYARROW_WITH_GANDIVA=0
+export PYARROW_WITH_PLASMA=1
+export PYARROW_WITH_S3=0
+export PYARROW_CMAKE_GENERATOR=Ninja
+BUILD_EXT_FLAGS=""
 
-# We'll add CUDA support later
-export PYARROW_WITH_CUDA=0
+# Enable CUDA support
+if [ "${build_type}" = "cuda" ]; then
+    export PYARROW_WITH_CUDA=1
+else
+    export PYARROW_WITH_CUDA=0
+fi
 
 cd python
 
@@ -43,4 +49,6 @@ $PYTHON setup.py \
         install --single-version-externally-managed \
                 --record=record.txt
 
-rm -r ${SP_DIR}/pyarrow/tests
+if [[ "$PKG_NAME" == "pyarrow" ]]; then
+    rm -r ${SP_DIR}/pyarrow/tests
+fi
